requirements.txt
streamlit
pandas
mysql-connector-python==8.0.33
